Moving _millis and _total_seconds methods from bigquery to core.
